News App
This project shows the simplest way of creating a customized login. With this simple android app those with the password and username can easily login to the app and upload there messages. The app lists the message in the order they were sent i.e the last news uploaded comes first in the application